explicitly allow only certain keys to be record
Yesterday modified the data-saving so that, for free-form text entry,
the final result is saved (not the data stream, which would include
control characters such as backspace).

Also, added quotations around the free-form response data (which might
include commas) so that the result should always be "valid" CSV.

To keep the data "valid" CSV, filtered for "newline" characters, to
prevent the CSV being corrupted by line breaks.

Also, explicitly allowed only the keys below to be received. A couple
more minor updates to the software to follow.